@@ -7,22 +7,22 @@ class OmniauthFinishedController < ApplicationController
     @user = User.find(params[:id])
     @provider = params[:provider]
 
-    if (request.post? || request.patch?) && @user.update(user_params)
-      # OAuth の場合には希望アカウントが入力された後にグループを作る
-      @user.create_default_group
-      @user.use_gravatar_icon(false)
-      # @user.send_confirmation_instructions unless @user.confirmed?
-      sign_in(@user, bypass: true)
-      redirect_to root_url, notice: t('devise.omniauth_callbacks.success', kind: @provider.capitalize)
-    end
+    return unless (request.post? || request.patch?) && @user.update(user_params)
+
+    # OAuth の場合には希望アカウントが入力された後にグループを作る
+    @user.create_default_group
+    # @user.use_gravatar_icon(false)
+    # @user.send_confirmation_instructions unless @user.confirmed?
+    sign_in(@user, bypass: true)
+    redirect_to root_url, notice: t('devise.omniauth_callbacks.success', kind: @provider.capitalize)
   end
 
   private
 
   # user_paramsにアクセスするため。
   def user_params
-    accessible = [ :username, :email ]
-    accessible << [ :password, :password_confirmation ] unless params[:user][:password].blank?
+    accessible = %i[username email]
+    accessible << %i[password password_confirmation] unless params[:user][:password].blank?
     params.require(:user).permit(accessible)
   end
 end
