@@ -86,7 +86,7 @@ export default {
 <style lang="scss" scoped>
 $vs-border-width: 0px !default;
 
-/deep/ {
+::v-deep {
   @import "~vue-select/src/scss/vue-select.scss";
 }
 
@@ -94,7 +94,7 @@ $vs-border-width: 0px !default;
   border-bottom: 1px solid #888
 }
 
-/deep/ .vs__search::placeholder {
+::v-deep .vs__search::placeholder {
   color: #b7b9cc;
 }
 </style>
