@@ -1,7 +1,7 @@
 Rails.application.routes.draw do
   devise_for :users, controllers: { omniauth_callbacks: 'omniauth_callbacks' }
   # OmniAuth認証後、email入力を求める処理のため。
-  match '/users/:id/finish_signup' => 'omniauth_finished#finish_signup', via: [:get, :patch], as: :finish_signup
+  match '/users/:id/finish_signup' => 'omniauth_finished#finish_signup', via: %i[get patch], as: :finish_signup
 
   # For details on the DSL available within this file, see https://guides.rubyonrails.org/routing.html
 
@@ -42,7 +42,7 @@ Rails.application.routes.draw do
   get '/profiles', to: 'profiles#index'
   patch '/profiles', to: 'profiles#update'
   patch '/profiles/update_password', to: 'profiles#update_password'
-  patch '/profiles/use_gravatar', to: 'profiles#use_gravatar'
+  # patch '/profiles/use_gravatar', to: 'profiles#use_gravatar'
   delete '/profiles/image', to: 'profiles#destroy_image'
 
   namespace :application_settings do
@@ -51,16 +51,18 @@ Rails.application.routes.draw do
     post '/', to: 'dashboard#update'
 
     patch 'group_managements/:id/add_user', to: 'group_managements#add_user', as: 'group_managements_add_user'
-    delete 'group_managements/:id/remove_user/:user_id', to: 'group_managements#remove_user', as: 'group_managements_remove_user'
+    delete 'group_managements/:id/remove_user/:user_id', to: 'group_managements#remove_user',
+                                                         as: 'group_managements_remove_user'
     resources :group_managements
 
     patch 'user_managements/:id/update_role', to: 'user_managements#update_role', as: 'user_managements_update_role'
-    delete 'user_managements/:id/destroy_image', to: 'user_managements#destroy_image', as: 'user_managements_destroy_image'
+    delete 'user_managements/:id/destroy_image', to: 'user_managements#destroy_image',
+                                                 as: 'user_managements_destroy_image'
     resources :user_managements
   end
 
   get '/installer', to: 'installer#index', as: 'installer'
-  match '/installer', to: 'installer#update', via: [:post, :patch]
+  match '/installer', to: 'installer#update', via: %i[post patch]
 
   get '/health', to: 'health#health'
 
